+++ problems/problem_31.hpp
@@ -3,6 +3,6 @@ namespace Computing
 {
     using namespace std;
 
-    void Combine(set<int> & coins, set<int>::iterator it, int val, int & count);
+    void Combine(set<int> & coins, set<int>::iterator it, int val, long long & count);
 }
 int problem_31();
